+++ test/typeparam/issue51219.out
@@ -1,1 +1,1 @@
-{} {{}}
+{}
