+++ test/typeparam/issue51219b.go
@@ -1,11 +1,7 @@
+// compiledir -G=3
+
 // Copyright 2022 The Go Authors. All rights reserved.
 // Use of this source code is governed by a BSD-style
 // license that can be found in the LICENSE file.
 
-package b
-
-import "a"
-
-type InteractionRequest[T a.InteractionDataConstraint] struct {
-	a.Interaction[T]
-}
+package ignored
